// dashboard/src/utils/hasura-api/generated/schemas/inputArgument.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */

export interface InputArgument {
  /** Name of the argument */
  name: string;
  /** GraphQL type of the argument */
  type: string;
}